+++ core/src/PrjFileProc.ts
@@ -13,7 +13,7 @@ import {ExtensionContext, workspace, Disposable, tasks, Task, ShellExecution, wi
 const fs = require('fs-extra');
 const path = require('path');
 const img_size = require('image-size');
-const crypt = require('crypto-js');
+const crypto = require('crypto-js');
 const uuidv4 = require('uuid/v4');
 const crc32 = require('crc-32');
 import {Transform} from 'stream';
@@ -27,16 +27,15 @@ export class PrjFileProc {
 	private	readonly	curPlg	: string;
 	private	readonly	curPrj	: string;
 	private	readonly	lenCurPrj: number;
-	private	readonly	curCrypt: string;
-	private readonly	fld_crypt_prj	= 'crypt_prj';
-//	private readonly	fld_crypt_prj	= '.prj';
-	private		$isCryptMode	= true;
-	get isCryptMode() {return this.$isCryptMode;}
-	private	regNeedCrypt	= /\.(sn|json|jpe?g|png|svg|webp|mp3|m4a|ogg|aac|flac|wav|mp4|webm|ogv)$/;
-	private	regFullCrypt	= /\.(sn|json)$/;
+	private	readonly	curCrypto: string;
+	private readonly	fld_crypto_prj	= 'crypto_prj';
+//	private readonly	fld_crypto_prj	= '.prj';
+	private		$isCryptoMode	= true;
+	get isCryptoMode() {return this.$isCryptoMode;}
+	private	regNeedCrypto	= /\.(sn|json|html?|jpe?g|png|svg|webp|mp3|m4a|ogg|aac|flac|wav|mp4|webm|ogv|html?)$/;
+	private	regFullCrypto	= /\.(sn|json|html?)$/;
 	private	regRepJson		= /\.(jpe?g|png|svg|webp|mp3|m4a|ogg|aac|flac|wav|mp4|webm|ogv)"/g;
 		// この末端の「"」は必須。変更時は delPrj_sub() 内も
-	private	regForceCrypt	= /\.(sn)$/;
 	private readonly	hExt2N: {[name: string]: number} = {
 		'jpg'	: 1,
 		'jpeg'	: 1,
@@ -50,9 +49,8 @@ export class PrjFileProc {
 		'flac'	: 14,
 		'wav'	: 15,
 		'mp4'	: 20,
-		'ogv'	: 21,
-		'webm'	: 22,
-		// woff2、otf、ttf
+		'webm'	: 21,
+		'ogv'	: 22,
 	};
 
 	private	readonly	hPass: {
@@ -114,25 +112,25 @@ export class PrjFileProc {
 			fwPrjJs.onDidChange(e=> this.chgPrj(e)),
 		];	// NOTE: ワークスペースだと、削除イベントしか発生しない？？
 
-		this.curCrypt = dir +`/${this.fld_crypt_prj}/`;
-		this.$isCryptMode = fs.existsSync(this.curCrypt);
+		this.curCrypto = dir +`/${this.fld_crypto_prj}/`;
+		this.$isCryptoMode = fs.existsSync(this.curCrypto);
 		const fnPass = this.curPlg +'/pass.json';
 		const exists_pass = fs.existsSync(fnPass);
 		this.hPass = exists_pass
 			? fs.readJsonSync(fnPass, {throws: false})
 			: {
 				pass	: uuidv4(),
-				salt	: String(crypt.lib.WordArray.random(128 / 8)),
-				iv		: String(crypt.lib.WordArray.random(128 / 8)),
+				salt	: String(crypto.lib.WordArray.random(128 / 8)),
+				iv		: String(crypto.lib.WordArray.random(128 / 8)),
 				ite		: 500 + Math.floor(new Date().getTime() %300),
-				stk		: String(crypt.lib.WordArray.random(128 / 8)),
+				stk		: String(crypto.lib.WordArray.random(128 / 8)),
 			};
 		if (! exists_pass) fs.outputJsonSync(fnPass, this.hPass);
 
-		this.iv = crypt.enc.Hex.parse(this.hPass.iv);
-		this.pbkdf2 = crypt.PBKDF2(
-			crypt.enc.Utf8.parse(this.hPass.pass),
-			crypt.enc.Hex.parse(this.hPass.salt),
+		this.iv = crypto.enc.Hex.parse(this.hPass.iv);
+		this.pbkdf2 = crypto.PBKDF2(
+			crypto.enc.Utf8.parse(this.hPass.pass),
+			crypto.enc.Hex.parse(this.hPass.salt),
 			{keySize: this.hPass.keySize, iterations: this.hPass.ite},
 		);
 
@@ -141,7 +139,7 @@ export class PrjFileProc {
 			this.hDiff = fs.readJsonSync(this.fnDiff);
 		}
 		this.ps = new PnlPrjSetting(ctx, dir, chgTitle);
-		if (this.$isCryptMode) this.initCrypt();
+		this.initCrypto();
 	}
 
 	private	ps: PnlPrjSetting;
@@ -150,13 +148,13 @@ export class PrjFileProc {
 	dispose() {this.aFSW.forEach(f=> f.dispose());}
 
 
-	private	crePrj(e: Uri) {this.chkAndEnc(e.path); this.updPathJson();}
-	private	chgPrj(e: Uri) {this.chkAndEnc(e.path);}
+	private	crePrj(e: Uri) {this.encIfNeeded(e.path); this.updPathJson();}
+	private	chgPrj(e: Uri) {this.encIfNeeded(e.path);}
 	private	delPrj(e: Uri) {
 		const short_path = e.path.slice(this.lenCurPrj);
-		this.regNeedCrypt.lastIndex = 0;
-		fs.removeSync(this.curCrypt
-			+ (short_path +'"')
+		this.regNeedCrypto.lastIndex = 0;
+		fs.removeSync(
+			this.curCrypto + (short_path +'"')
 			.replace(this.regRepJson, '.bin')
 			.replace(/"/, '')
 		);
@@ -166,17 +164,52 @@ export class PrjFileProc {
 		this.updDiffJson();
 	}
 
+	// プロジェクトフォルダ以下全走査で暗号化
+	private initCrypto() {
+		treeProc(
+			this.curPrj,
+			this.$isCryptoMode
+				? url=> {if (this.isDiff(url)) this.encrypter(url)}
+				: url=> this.isDiff(url)
+		);
+		this.updDiffJson();
+	}
+	private	encIfNeeded(url: string) {
+		if (this.isDiff(url) && this.$isCryptoMode) this.encrypter(url);
+		this.updDiffJson();
+	}
+	private	updDiffJson() {fs.writeJsonSync(this.fnDiff, this.hDiff);}
+	private	readonly	LEN_CHKDIFF		= 1024;
+	private	isDiff(url: string): boolean {
+		const short_path = url.slice(this.lenCurPrj);
+		let hash = 0;
+		if (this.regFullCrypto.test(url)) {
+			hash = crc32.str(fs.readFileSync(url, {encoding: 'utf8'}));
+		}
+		else {
+			const b = new Uint8Array(this.LEN_CHKDIFF);
+			const fd = fs.openSync(url, 'r');
+			fs.readSync(fd, b, 0, this.LEN_CHKDIFF, 0);
+			fs.closeSync(fd);
+			hash = crc32.buf(b);
+		}
+		if (this.hDiff[short_path] == hash) return false;
+
+		this.hDiff[short_path] = hash;
+		return true;
+	}
+
 
 	private	readonly	aRepl = [
 		'core/app4webpack.js',
 		'core/mob4webpack.js',
 		'core/web4webpack.js',
 	];
-	tglCryptMode() {
+	tglCryptoMode() {
 		const pathPre = this.curPlg +'/snsys_pre';
-		if (this.$isCryptMode) {
-			fs.removeSync(this.curCrypt);
-			this.$isCryptMode = false;
+		this.$isCryptoMode = ! this.$isCryptoMode;
+		if (! this.$isCryptoMode) {
+			fs.removeSync(this.curCrypto);
 
 			fs.removeSync(pathPre);
 
@@ -190,26 +223,27 @@ export class PrjFileProc {
 			// ビルド情報：パッケージするフォルダ名変更
 			replaceFile(
 				this.dir +'/package.json',
-				new RegExp(`"${this.fld_crypt_prj}\\/",`),
+				new RegExp(`"${this.fld_crypto_prj}\\/",`),
 				`"prj/",`,
 			);
+
 			return;
 		}
-		fs.ensureDir(this.curCrypt);
-		this.$isCryptMode = true;
+
+		fs.ensureDir(this.curCrypto);
 
 		// SKYNovelが見に行くプロジェクトフォルダ名変更
 		this.aRepl.forEach(url=> replaceFile(
 			this.dir +'/'+ url,
 			/\(hPlg\);/,
-			`(hPlg, {cur: '${this.fld_crypt_prj}/', crypt: true});`,
+			`(hPlg, {cur: '${this.fld_crypto_prj}/', crypto: true});`,
 		));
 
 		// ビルド情報：パッケージするフォルダ名変更
 		replaceFile(
 			this.dir +'/package.json',
 			/"prj\/",/,
-			`"${this.fld_crypt_prj}/",`,
+			`"${this.fld_crypto_prj}/",`,
 		);
 
 		// プラグインソースに埋め込む
@@ -220,75 +254,40 @@ export class PrjFileProc {
 			pathPre +'/index.js',
 		);
 
-		this.initCrypt();
+		this.hDiff = Object.create(null);
+		this.initCrypto();
 	}
-
-	// プロジェクトフォルダ以下全走査で暗号化
-	private initCrypt() {
-		treeProc(this.curPrj, url=>{if (this.isDiff(url)) this.encrypter(url)});
-		this.updDiffJson();
-	}
-	private	chkAndEnc(url: string) {
-		if (! this.isDiff(url)) return;
-		this.encrypter(url);
-		this.updDiffJson();
-	}
-	private	readonly	LEN_CHKDIFF		= 1024;
-	private	isDiff(url: string): boolean {
-		const short_path = url.slice(this.lenCurPrj);
-		let hash = 0;
-		if (this.regFullCrypt.test(url)) {
-			if (short_path == 'path.json') return false;	// updPathJson()任せ
-			hash = crc32.str(fs.readFileSync(url, {encoding: 'utf8'}));
-		}
-		else {
-			const b = new Uint8Array(this.LEN_CHKDIFF);
-			const fd = fs.openSync(url, 'r');
-			fs.readSync(fd, b, 0, this.LEN_CHKDIFF, 0);
-			fs.closeSync(fd);
-			hash = crc32.buf(b);
-		}
-		const isDiff = (this.hDiff[short_path] != hash)
-		if (! isDiff) return false;
-
-		this.hDiff[short_path] = hash;
-		return true;
-	}
-	private	updDiffJson() {fs.writeJsonSync(this.fnDiff, this.hDiff);}
 
 	private	pbkdf2	: any;
 	private	iv		: any;
 	private	readonly LEN_ENC	= 1024 *10;
 	private	readonly regDir = /(^.+)\//;
 	private	async encrypter(url: string) {
-		if (! this.$isCryptMode) return;
-
 		try {
 			const short_path = url.slice(this.lenCurPrj);
-			const url_out = this.curCrypt + short_path;
-			if (! this.regNeedCrypt.test(url)) {
+			const url_out = this.curCrypto + short_path;
+			if (! this.regNeedCrypto.test(url)) {
 				fs.ensureLink(url, url_out)
 				.catch((e: any)=> console.error(`encrypter cp1 ${e}`));
 				return;
 			}
 
-			if (! this.regForceCrypt.test(url)) {
+			if (this.regFullCrypto.test(short_path)) {
+				let s = await fs.readFile(url, {encoding: 'utf8'});
+				if (short_path == 'path.json') {	// 内容も変更
+					s = s.replace(this.regRepJson, '.bin"');
+				}
+				const e = crypto.AES.encrypt(s, this.pbkdf2, {iv: this.iv});
+				await fs.outputFile(url_out, e.toString());
+				return;
+			}
+			else {
 				const dir = this.regDir.exec(short_path);
 				if (dir && this.ps.cfg.code[dir[1]]) {
 					fs.ensureLink(url, url_out)
 					.catch((e: any)=> console.error(`encrypter cp2 ${e}`));
 					return;
 				}
-			}
-
-			if (this.regFullCrypt.test(short_path)) {
-				let s = await fs.readFile(url, {encoding: 'utf8'});
-				if (short_path == 'path.json') {	// 内容も変更
-					s = s.replace(this.regRepJson, '.bin"');
-				}
-				const e = crypt.AES.encrypt(s, this.pbkdf2, {iv: this.iv});
-				await fs.outputFile(url_out, e.toString());
-				return;
 			}
 
 			let nokori = this.LEN_ENC;
@@ -300,7 +299,7 @@ export class PrjFileProc {
 			const rs = fs.createReadStream(url)
 			.on('error', (e :any)=> console.error(`encrypter rs=%o`, e));
 
-			const u2 = url_out.replace(/\..+$/, '\.bin');
+			const u2 = url_out.replace(/\..+$/, '.bin');
 			fs.ensureFileSync(u2);	// touch
 			const ws = fs.createWriteStream(u2)
 			.on('error', (e :any)=> console.error(`encrypter ws=%o`, e));
@@ -318,8 +317,8 @@ export class PrjFileProc {
 				}
 
 				bh.set(chunk.slice(0, nokori), i);
-				const e6 = crypt.AES.encrypt(
-					crypt.lib.WordArray.create(bh),
+				const e6 = crypto.AES.encrypt(
+					crypto.lib.WordArray.create(bh),
 					this.pbkdf2,
 					{iv: this.iv},
 				);
@@ -337,8 +336,8 @@ export class PrjFileProc {
 			.on('end', ()=> {
 				if (nokori == 0) return;
 
-				const e6 = crypt.AES.encrypt(
-					crypt.lib.WordArray.create(bh.slice(0, i)),
+				const e6 = crypto.AES.encrypt(
+					crypto.lib.WordArray.create(bh.slice(0, i)),
 					this.pbkdf2,
 					{iv: this.iv},
 				);
@@ -384,7 +383,7 @@ export class PrjFileProc {
 		try {
 			const hPath = this.get_hPathFn2Exts(this.curPrj);
 			await fs.outputJson(this.curPrj +'path.json', hPath);
-			this.encrypter(this.curPrj +'path.json');
+			if (this.$isCryptoMode) this.encrypter(this.curPrj +'path.json');
 		}
 		catch (err) {console.error(`PrjFileProc updPathJson ${err}`);}
 	}